requirements.txt
djangorestframework==3.12.2
pytesseract
opencv-python
django-cors-headers==3.1.0
Django==3.1.7
psycopg2
celery[redis]
pytest
pytest-django
pytest-mock